PMI-503: Restructure project for better organisation. Move upload executable script to `\bin` directory. Move business logic into `performance` package. Move utility functions to `utils` package. Co-authored-by: George Lund <george.lund@digital.cabinet-office.gov.uk>
