+++ bin/bootstrap.py
@@ -1,0 +1,5 @@
+import sys
+from dotenv import load_dotenv
+
+sys.path.append('../')
+load_dotenv(override=True)
