+++ bin/upload.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+
+"""
+Upload a file to S3.
+
+Usage:
+    python3 upload.py -f <file-name>
+"""
+
+import bootstrap
+import argparse
+from utils.env import check_get_env
+from utils import aws
+from performance.uploader import Uploader
+
+RP_REPORT_OUTPUT_BUCKET = check_get_env('RP_REPORT_OUTPUT_BUCKET')
+
+
+def main(file_name):
+    resource = aws.get_s3_resource()
+    uploader = Uploader(resource)
+    uploader.upload(RP_REPORT_OUTPUT_BUCKET, file_name)
+    print('File uploaded successfully.')
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser()
+    parser.add_argument('filename', help='File to upload', type=str)
+    args = parser.parse_args()
+    main(args.filename)
